feat: use hover for inFiles problem navigation
I was surprised to find out that despite the native navigation being used,
the previous implementation still was opening the next/prev files manually
before it. This lead to inconsistencies since the first time a file would
be navigated, it would display a native marker but then continue with
hover diagnostics.

So I fixed that by not using the native commands but rather invoking hover
on already identified diagnostics.